fix also the other EIC projections
