--- nnpdf_data/nnpdf_data/new_commondata/ATHENA_NC_140GEV_EP/athena_140gev_ep
@@ -1,131 +1,0 @@
-0.000129	1.29
-0.000205	1.29
-0.000205	2.05
-0.000325	1.29
-0.000325	2.05
-0.000325	3.25
-0.000515	1.29
-0.000515	2.05
-0.000515	3.25
-0.000515	5.15
-0.000815	1.29
-0.000815	2.05
-0.000815	3.25
-0.000815	5.15
-0.000815	8.15
-0.00129	1.29
-0.00129	2.05
-0.00129	3.25
-0.00129	5.15
-0.00129	8.15
-0.00129	12.9
-0.00205	1.29
-0.00205	2.05
-0.00205	3.25
-0.00205	5.15
-0.00205	8.15
-0.00205	12.9
-0.00205	20.5
-0.00325	1.29
-0.00325	2.05
-0.00325	3.25
-0.00325	5.15
-0.00325	8.15
-0.00325	12.9
-0.00325	20.5
-0.00325	32.5
-0.00515	2.05
-0.00515	3.25
-0.00515	5.15
-0.00515	8.15
-0.00515	12.9
-0.00515	20.5
-0.00515	32.5
-0.00515	51.5
-0.00815	3.25
-0.00815	5.15
-0.00815	8.15
-0.00815	12.9
-0.00815	20.5
-0.00815	32.5
-0.00815	51.5
-0.00815	81.5
-0.0129	5.15
-0.0129	8.15
-0.0129	12.9
-0.0129	20.5
-0.0129	32.5
-0.0129	51.5
-0.0129	81.5
-0.0129	129
-0.0205	8.15
-0.0205	12.9
-0.0205	20.5
-0.0205	32.5
-0.0205	51.5
-0.0205	81.5
-0.0205	129
-0.0205	205
-0.0325	12.9
-0.0325	20.5
-0.0325	32.5
-0.0325	51.5
-0.0325	81.5
-0.0325	129
-0.0325	205
-0.0325	325
-0.0515	20.5
-0.0515	32.5
-0.0515	51.5
-0.0515	81.5
-0.0515	129
-0.0515	205
-0.0515	325
-0.0515	515
-0.0815	32.5
-0.0815	51.5
-0.0815	81.5
-0.0815	129
-0.0815	205
-0.0815	325
-0.0815	515
-0.0815	815
-0.129	51.5
-0.129	81.5
-0.129	129
-0.129	205
-0.129	325
-0.129	515
-0.129	815
-0.129	1.29E+03
-0.205	81.5
-0.205	129
-0.205	205
-0.205	325
-0.205	515
-0.205	815
-0.205	1.29E+03
-0.205	2.05E+03
-0.325	129
-0.325	205
-0.325	325
-0.325	515
-0.325	815
-0.325	1.29E+03
-0.325	2.05E+03
-0.325	3.25E+03
-0.515	205
-0.515	325
-0.515	515
-0.515	815
-0.515	1.29E+03
-0.515	2.05E+03
-0.515	3.25E+03
-0.515	5.15E+03
-0.815	325
-0.815	515
-0.815	815
-0.815	1.29E+03
-0.815	2.05E+03
-0.815	3.25E+03
-0.815	5.15E+03
