+++ nnpdf_data/nnpdf_data/new_commondata/ATHENA_NC_45GEV_EP/rawdata/ATHENA_NC_45GEV_EP.yaml
