+++ nnpdf_data/nnpdf_data/new_commondata/ATHENA_NC_63GEV_EP/rawdata/ATHENA_NC_63GEV_EP.yaml
