+++ nnpdf_data/nnpdf_data/new_commondata/EIC_NC_211GEV_EP/rawdata/EIC_18_275_A1c_100fb-1.txt
