+++ nnpdf_data/nnpdf_data/new_commondata/EIC_NC_43GEV_EP/rawdata/EIC_5_41_A1c_100fb-1.txt
