+++ nnpdf_data/nnpdf_data/new_commondata/EIC_NC_43GEV_EP/rawdata/EIC_NC_43GEV_EP.yaml
