+++ nnpdf_data/nnpdf_data/new_commondata/EIC_NC_67GEV_EP/rawdata/EIC_5_100_A1c_100fb-1.txt
