+++ nnpdf_data/nnpdf_data/new_commondata/EIcC_NC_15GEV_EP/rawdata/EIcC_3_20_A1c_100fb-1.txt
