+++ nnpdf_data/nnpdf_data/new_commondata/EIcC_NC_15GEV_EP/rawdata/EIcC_NC_15GEV_EP.yaml
