+++ nnpdf_data/nnpdf_data/new_commondata/EIcC_NC_22GEV_EP/rawdata/EIcC_5_25_A1c_100fb-1.txt
