+++ nnpdf_data/nnpdf_data/new_commondata/EIcC_NC_22GEV_EP/rawdata/EIcC_NC_22GEV_EP.yaml
